Fixed import names and split up audio controls stylesheet
