+++ src/components/audio-player/index.js
@@ -1,12 +1,10 @@
-import AudioControls from './AudioControls';
-import AudioVisualisation from './AudioVisualisation';
-import AudioScrubber from './AudioScrubber';
-import AudioPlayerCore from './AudioPlayerCore';
+import AudioControls from '../../containers/audio-controls';
+import AudioVisualisation from '../../containers/audio-visualisation';
+import AudioScrubber from '../../containers/audio-scrubber';
+import AudioPlayerCore from '../../containers/audio-player-core';
 
 import React from 'react';
 import ImmutableComponent from '../../ImmutableComponent';
-
-import './style.scss';
 
 export default class AudioPlayer extends ImmutableComponent {
     render() {
