+++ src/containers/audio-visualisation/style.scss
@@ -1,0 +1,15 @@
+.visualiser {
+    display: flex;
+    align-items: flex-end;
+    flex-grow: 1;
+    height: 56px;
+
+    background-image: linear-gradient(to bottom, #eee 1px, transparent 1px), linear-gradient(to right, #eee 1px, transparent 1px);
+    background-size: 5px 5px;
+
+    .visualiser-canvas {
+        width: 100%;
+        height: inherit;
+    }
+}
+
