+++ src/index.js
@@ -7,7 +7,7 @@ import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 
 import getStore from './store';
-import App from './components/App';
+import App from './components/app';
 
 import './images/favicon.png';
